Backend/requirements.txt
astroid==2.2.5
backports.csv==1.0.7
beautifulsoup4==4.8.0
bs4==0.0.1
captcha-solver==0.1.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==2.4.0
django-filter==2.1.0
django-import-export==1.2.0
django-rest-multiple-models==2.1.3
djangorestframework==3.9.1
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.1
lxml==4.4.0
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
pylint==2.3.1
pytesseract==0.2.7
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tablib==0.13.0
twocaptchaapi==0.3
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
